🚚 Move shadcn utils to under utils folder
