+++ client/components.json
@@ -12,7 +12,7 @@
   },
   "aliases": {
     "components": "@/components",
-    "utils": "@/lib/utils",
+    "utils": "@/lib/utils/shadcn",
     "ui": "@/components/atoms",
     "lib": "@/lib",
     "hooks": "@/hooks"
