+++ client/src/components/atoms/sidebar.tsx
@@ -3,7 +3,7 @@ import { Slot } from "@radix-ui/react-slot"
 import { VariantProps, cva } from "class-variance-authority"
 import { PanelLeftIcon } from "lucide-react"
 import { useIsMobile } from "@/lib/hooks/use-mobile"
-import { cn } from "@/lib/utils"
+import { cn } from "@/lib/utils/shadcn"
 import { Button } from "@/components/atoms/button"
 import { Input } from "@/components/atoms/input"
 import { Separator } from "@/components/atoms/separator"
