+++ client/src/components/atoms/skeleton.tsx
@@ -1,4 +1,4 @@
-import { cn } from "@/lib/utils"
+import { cn } from "@/lib/utils/shadcn"
 
 function Skeleton({ className, ...props }: React.ComponentProps<"div">) {
   return (
